remove domain dir
